* 'master' of https://github.com/scikit-learn/scikit-learn: (283 commits)
  update file header in affinity_propagation.py
  DOC: Fix spelling
  adding 0.17 support link
  adding missing 0.16 doc link and removing broken support links
  Fixed unit test in sklearn/tree/tests/test_tree.py.
  fix link to contributing in website drop-down
  Fix Kernel PCA docstring to reflect how remove_zero_eig defaults to false
  removed the doc line 'Only applies if analyzer == 'word'' from CountVectorizer analyzer'
  Parallelize embarrassingly parallel loop in RFECV.fit
  Changed orders of docstring descriptions for polynomial_kernel
  Added more gamma info for sigmoid, polynomial kernels
  Added default gamma info for laplacian_kernel and rbf_kernel
  Add link to multiisotonic
  ENH add multinomial SAG solver for LogisticRegression
  change comment, printed comment, and code behavior to all reflect computation of mean squared error
  residual sum of squares should be computed by summing the squared differences instead of taking the mean.
  repaired dead link
  It's better to fix a comment
  Fix coveralls report sending
  edited docstring for clarity regarding the naming of the pipeline components.
  ...

+++ sklearn/manifold/tests/test_t_sne.py
@@ -143,7 +143,7 @@ def test_binary_search_neighbors():
     for k in np.linspace(80, n_samples, 10):
         k = int(k)
         topn = k * 10  # check the top 10 *k entries out of k * k entries
-        neighbors_nn = np.argsort(distances, axis=1)[:, :k]
+        neighbors_nn = np.argsort(distances, axis=1)[:, :k].astype(np.int64)
         P2k = _binary_search_perplexity(distances, neighbors_nn,
                                         desired_perplexity, verbose=0)
         idx = np.argsort(P1.ravel())[::-1]
@@ -164,7 +164,7 @@ def test_binary_perplexity_stability():
     distances = np.abs(distances.dot(distances.T))
     np.fill_diagonal(distances, 0.0)
     last_P = None
-    neighbors_nn = np.argsort(distances, axis=1)[:, :k]
+    neighbors_nn = np.argsort(distances, axis=1)[:, :k].astype(np.int64)
     for _ in range(100):
         P = _binary_search_perplexity(distances.copy(), neighbors_nn.copy(),
                                       3, verbose=0)
